restfulapicrud/requirements.txt
kafka-python==2.0.2
Django==3.1.7
djangorestframework==3.12.4
django-cors-headers==3.7.0
psycopg2
